services/scraper/requirements.txt
anthropic>=0.52.0
playwright>=1.44.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
supabase>=2.4.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.25.0
loguru>=0.7.0
